feat: Implemented new action bar, made years scroll horizontally, split user entry logic away from jsx
