+++ src/index.css
@@ -25,6 +25,7 @@ code {
 
 /* width */
 ::-webkit-scrollbar {
+  height:10px;
   width: 8px;
 }
